:white_check_mark: Bring changes through test files

+++ test/IdeaTokenHub.t.sol
@@ -244,7 +244,7 @@ contract IdeaTokenHubTest is NounsEnvSetup, TestUtils {
             // reduce an entropic hash to the `[0:nextIdeaId]` range via modulo
             uint256 numIds = ideaTokenHub.getNextIdeaId() - 1;
             // add 1 since modulo produces one less than desired range, incl 0
-            uint256 pseudoRandomIdeaId = (uint256(keccak256(abi.encode(l))) % numIds) + 1;
+            uint96 pseudoRandomIdeaId = uint96((uint256(keccak256(abi.encode(l))) % numIds)) + 1;
             uint256 currentIdTotalFunding = ideaTokenHub.getIdeaInfo(pseudoRandomIdeaId).totalFunding; // get existing funding value
 
             vm.expectEmit(true, true, true, false);
@@ -399,7 +399,7 @@ contract IdeaTokenHubTest is NounsEnvSetup, TestUtils {
             // reduce an entropic hash to the `[0:nextIdeaId]` range via modulo
             uint256 numIds = ideaTokenHub.getNextIdeaId() - 1;
             // add 1 since modulo produces one less than desired range, incl 0
-            uint256 pseudoRandomIdeaId = (uint256(keccak256(abi.encode(l))) % numIds) + 1;
+            uint96 pseudoRandomIdeaId = uint96(uint256(keccak256(abi.encode(l))) % numIds) + 1;
             uint256 currentIdTotalFunding = ideaTokenHub.getIdeaInfo(pseudoRandomIdeaId).totalFunding; // get existing funding value
 
             vm.expectEmit(true, true, true, false);
@@ -585,7 +585,7 @@ contract IdeaTokenHubTest is NounsEnvSetup, TestUtils {
             // reduce an entropic hash to the `[0:nextIdeaId]` range via modulo
             uint256 numIds = ideaTokenHub.getNextIdeaId() - 1;
             // add 1 since modulo produces one less than desired range, incl 0
-            uint256 pseudoRandomIdeaId = (uint256(keccak256(abi.encode(l))) % numIds) + 1;
+            uint96 pseudoRandomIdeaId = uint96((uint256(keccak256(abi.encode(l))) % numIds)) + 1;
             uint256 currentIdTotalFunding = ideaTokenHub.getIdeaInfo(pseudoRandomIdeaId).totalFunding; // get existing funding value
 
             vm.expectEmit(true, true, true, false);
